data_collection/data_collection/requirements.txt
requests
pymongo
python-dotenv
pytrends
textblob
